Fix rubocop issues and refactor spec support files
- fix: travis build
- fix: mute console output only runs on tagged example groups
- refactor: reorganize/rename several support files
