+++ spec/support/mute_console_output.rb
@@ -3,13 +3,13 @@ RSpec.configure do |config|
   original_stderr = $stderr
   original_stdout = $stdout
 
-  config.before(:all) do
+  config.before(:context, mute_output: true) do
     # Redirect stderr and stdout
     $stderr = File.open(File::NULL, 'w')
     $stdout = File.open(File::NULL, 'w')
   end
 
-  config.after(:all) do
+  config.after(:context, mute_output: true) do
     [$stderr, $stdout].each(&:close)
     $stderr = original_stderr
     $stdout = original_stdout
